+++ queries/gdshader/imports.scm
@@ -1,0 +1,1 @@
+(preproc_include path: (string_literal (string_content) @module))